requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
black==21.10b0
celery==5.1.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-stripe==2.5.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonfield==3.1.0
kombu==5.2.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient @ file:///C:/Users/user/Documents/mysqlclient-2.0.3-cp39-cp39-win_amd64.whl
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
PyJWT==1.7.1
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stripe==2.63.0
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
mysqlclient~=2.0.3